requirements.txt
bleach==3.1.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.4
djangorestframework==3.9.4
Markdown==3.1.1
Pillow==6.0.0
python-decouple==3.1
pytz==2018.7
six==1.12.0
static3==0.7.0
Unipath==1.1
webencodings==0.5.1
